Add nix configurations
